+++ Changelog.md
@@ -1,5 +1,18 @@
 # Changelog v1.8
 
 ## Additions
-- Added crimson moss
-- Added crimson-mossy blocks
+- Added blocks
+  - Added "Crimson Moss"
+  - Added "Crimson Moss Carpet"
+  - Added "Warped Moss"
+  - Added "Warped Moss Carpet"
+- Added items
+  - Added "Crimson Moss"
+  - Added "Crimson Moss Carpet"
+  - Added "Crimson Moss Ball"   
+  - Added "Warped Moss"
+  - Added "Warped Moss Carpet"
+  - Added "Warped Moss Ball"
+- Added new biomes
+  - "Crimson Moss Forest"
+  - "Warped Moss Forest"